solving bug for selected category
