+++ lib/widgets/expense_report_dialog.dart
@@ -1,8 +1,9 @@
-import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import '../models/expense.dart';
 import '../services/expense_service.dart';
-import '../views/attendant_expenses_view.dart';
+import 'category_dropdown.dart';
+import 'upload_document_button.dart';
+import 'all_expenses_button.dart'; // Import the AllExpenses widget
 
 class ExpenseReportDialog extends StatefulWidget {
   final List<Expense> expenses;
@@ -23,7 +24,7 @@ class ExpenseReportDialog extends StatefulWidget {
 
 class _ExpenseReportDialogState extends State<ExpenseReportDialog> {
   List<Expense> expenses = [];
-  String? selectedCategory; // Moved to a higher scope
+  String? selectedCategory;
 
   @override
   void initState() {
@@ -48,14 +49,18 @@ class _ExpenseReportDialogState extends State<ExpenseReportDialog> {
         IconButton(
           icon: Icon(Icons.edit_note, size: 24),
           onPressed: () async {
+            print('Before showDialog - selectedCategory: $selectedCategory');
             await widget
                 .onDialogOpened(); // Call the onDialogOpened callback here
             await showExpenseReportDialog(context, expenses, widget.buildingId,
                 _loadExpenses, selectedCategory, (String? newCategory) {
               setState(() {
                 selectedCategory = newCategory;
+                print(
+                    'Inside showDialog callback - selectedCategory updated to: $selectedCategory');
               });
             });
+            print('After showDialog - selectedCategory: $selectedCategory');
             // After the dialog is closed, you might want to refresh the expenses again
             // to ensure any changes made in the dialog are reflected.
             await _loadExpenses();
@@ -67,112 +72,7 @@ class _ExpenseReportDialogState extends State<ExpenseReportDialog> {
   }
 }
 
-class ExpenseSummary extends StatelessWidget {
-  final List<Expense> expenses;
-  final String buildingId;
-  final Future<void> Function() onExpensesUpdated;
-
-  const ExpenseSummary({
-    Key? key,
-    required this.expenses,
-    required this.buildingId,
-    required this.onExpensesUpdated,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Tooltip(
-      message: 'כל התשלומים',
-      child: GestureDetector(
-        onTap: () async {
-          final result = await showDialog(
-            context: context,
-            builder: (BuildContext context) {
-              return AttendantExpensesView(
-                buildingId: buildingId,
-              );
-            },
-          );
-          if (result == true) {
-            await onExpensesUpdated(); // Use the callback here
-          }
-          // Trigger state update in parent widget if needed
-        },
-        child:
-            Text('סך ההוצאות: ₪${expenses.fold(0.0, (a, b) => a + b.amount)}'),
-      ),
-    );
-  }
-}
-
-Widget _buildCategoryDropdown({
-  required List<String> categories,
-  required String? selectedCategory,
-  required StateSetter setState,
-  required BuildContext context,
-  required Function(String?) onSelectCategory,
-}) {
-  return DropdownButtonFormField<String>(
-    value: selectedCategory,
-    hint: Text('בחר קטגוריה'),
-    onChanged: (String? newValue) {
-      if (newValue == 'אחר...') {
-        showDialog<String>(
-          context: context,
-          builder: (BuildContext context) {
-            return AlertDialog(
-              title: Text('הזן את הקטגוריה'),
-              content: TextField(
-                autofocus: true,
-                onSubmitted: (String value) {
-                  Navigator.pop(context, value);
-                },
-              ),
-            );
-          },
-        ).then((String? value) {
-          if (value != null) {
-            setState(() {
-              categories.add(value);
-              selectedCategory = value;
-            });
-            onSelectCategory(value); // Use the callback to update
-          }
-        });
-      } else {
-        onSelectCategory(newValue); // Use the callback to update
-      }
-    },
-    items: categories.map<DropdownMenuItem<String>>((String value) {
-      return DropdownMenuItem<String>(
-        value: value,
-        child: Text(value),
-      );
-    }).toList(),
-  );
-}
-
-Widget _buildUploadDocumentButton({
-  required StateSetter setState,
-  required String? filePath,
-}) {
-  return TextButton.icon(
-    icon: Icon(Icons.upload_file),
-    label: Text('העלאת מסמך'),
-    onPressed: () async {
-      final result = await FilePicker.platform.pickFiles();
-      if (result != null && result.files.single.path != null) {
-        setState(() {
-          filePath = result.files.single.path;
-        });
-      } else {
-        // User canceled the picker or the operation failed
-      }
-    },
-  );
-}
-
-Future<void> showExpenseReportDialog(
+Future<String?> showExpenseReportDialog(
     BuildContext context,
     List<Expense> expenses,
     String buildingId,
@@ -189,7 +89,7 @@ Future<void> showExpenseReportDialog(
     'גינון',
     'אינסטלטור',
     'אחר...',
-  ]; // Example categories
+  ];
   await showDialog<void>(
     context: context,
     builder: (BuildContext context) {
@@ -205,16 +105,14 @@ Future<void> showExpenseReportDialog(
             return Column(
               mainAxisSize: MainAxisSize.min,
               children: [
-                ExpenseSummary(
+                AllExpensesButton(
                   expenses: expenses,
                   buildingId: currentBuildingId,
                   onExpensesUpdated: onExpensesUpdated,
                 ),
-                _buildCategoryDropdown(
+                CategoryDropdown(
                   categories: categories,
                   selectedCategory: selectedCategory,
-                  setState: setState,
-                  context: context,
                   onSelectCategory: onSelectCategory,
                 ),
                 TextFormField(
@@ -224,8 +122,13 @@ Future<void> showExpenseReportDialog(
                     labelText: 'סכום',
                   ),
                 ),
-                _buildUploadDocumentButton(
-                    setState: setState, filePath: filePath),
+                UploadDocumentButton(
+                  onFilePicked: (String? path) {
+                    setState(() {
+                      filePath = path;
+                    });
+                  },
+                ),
                 TextFormField(
                   controller: noteController,
                   decoration: InputDecoration(
@@ -243,13 +146,14 @@ Future<void> showExpenseReportDialog(
                       title: note,
                       amount: amount,
                       date: selectedDate,
-                      categoryId: selectedCategory ??
-                          'אחר', // Default to 'Other' if no category selected
+                      categoryId: selectedCategory ?? 'אחר',
                       filePath: filePath,
                     );
+                    print('newExpense.categoryId: ${newExpense.categoryId}');
                     ExpenseService expenseService = ExpenseService();
                     await expenseService.addExpense(
                         currentBuildingId, newExpense);
+                    setState(() {});
                     Navigator.of(context).pop(); // Close the dialog
                   },
                   child: Text('שלח'),
@@ -267,4 +171,5 @@ Future<void> showExpenseReportDialog(
       );
     },
   );
+  return selectedCategory; // Add this line before closing the dialog
 }
